+++ src/main/java/org/polyfrost/evergreenhud/mixins/client/Mixin_InventoryScreen_PlayerPreviewPartialTick.java
@@ -1,15 +1,14 @@
 package org.polyfrost.evergreenhud.mixins.client;
 
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import net.minecraft.client.gui.screens.inventory.InventoryScreen;
 import org.polyfrost.evergreenhud.client.hooks.PlayerPreviewPartialTick;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.Constant;
-import org.spongepowered.asm.mixin.injection.ModifyConstant;
+import org.spongepowered.asm.mixin.injection.At;
 
 @Mixin(InventoryScreen.class)
 public class Mixin_InventoryScreen_PlayerPreviewPartialTick {
-
-    @ModifyConstant(
+    @ModifyExpressionValue(
             //? if <1.21.4 {
             /*method = "method_29977",
             *///?} elif <1.21.8 {
@@ -19,11 +18,10 @@ public class Mixin_InventoryScreen_PlayerPreviewPartialTick {
             *///?} else {
             method = "extractRenderState(Lnet/minecraft/world/entity/LivingEntity;)Lnet/minecraft/client/renderer/entity/state/EntityRenderState;",
             //?}
-            constant = @Constant(floatValue = 1.0F)
+            at = @At(value = "CONSTANT", args = "floatValue=1.0")
     )
     private static float evergreenhud$playerPreviewPartialTick(float original) {
         float override = PlayerPreviewPartialTick.getPlayerPreviewPartialTick();
         return override < 0.0F ? original : override;
     }
-
 }