+++ CHANGELOG.md
@@ -1,2 +1,7 @@
-## 3.8.1
-- Fix coords hud
+## 3.9.0
+- Fixed Item Lore HUD
+- Fixed Keystrokes HUD outline
+- Added vanilla style for Inventory and Armor HUD
+- Improved performance of Position HUD
+- Added smoothing option to Speed HUD (enabled by default)
+- Added support for Minecraft 26.3